feat: Move multi tag config styles into global style config
BREAKING CHANGE

In order to make the select more branding friendly, i'm moving the style
source of truth into a configurable GlobalStyleConfig type. All the default
styling for the select will live here and be configurable by the
consumer.

For now I have added the multi tag styles to the
GlobalStyleConfiguration config and means
we no longer need a MultiSelectConfig type when using the multi variant.
It was only holding these styles so shouldn't be a huge deal to sort of
for consumers.