feat: Add placeholder option
The placeholder option is hidden and disabled so never registers as a selected option.

It does not show up in the option list.

When a selection happens this option does not appear again. We might
want to make it appear again if we add a delete option for native
select.