feat: Resolve index based on current selection
When there is no selection we render an extra select option that holds
the placeholder value. When we initially select an item when this extra
option exists we need to subtract it from the selected index that comes
from the "change" event. This is because if there is a placeholder
option, and a user opens and selects the first menu item, the changed
event says the 1 index has been selected which is technically true. We
however want to select the 0 index from the provided menu items to
select the correct item.

Once there is a selection then the index from the change event correctly
reflects the indicis from the provided menu items.