feat: Set InputChange when input is blurred
This will only happen if the input is changed from a non empty string to
an empty string on blur.

We want to avoid calling InputChange when a input value is already an
empty string as this may cause a consimer to over fetch something if for
example they are making api requests on input.