Focus when clearing Multi variant.
The VDOM removes focus form the input when more than one selected item
is being cleared. For safety, I am focusing the input again even though
it may already have focus to avoid being in a weird state where the
input is not focused, but the blur event was not dispatched.